requirements.txt
Flask
Flask-SQLAlchemy
Werkzeug
psycopg2
python-dotenv
opencensus-ext-azure
flask-socketio